requirements.txt
aiocoap
paho-mqtt

